requirements.txt
Flask==3.0.0
numpy==1.26.4
Pillow==10.2.0
requests==2.31.0

# Google API
google-api-python-client==2.149.0
google-auth==2.34.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0

# Face Recognition (requires dlib)
cmake==3.29.6
dlib==19.24.2
face-recognition==1.3.0
face-recognition-models==0.3.0

# OpenCV
opencv-python==4.9.0.80
